Work towards Windows build
Add make.bat for Windows build
Split main.cpp and file_io.cpp

+++ src/file_io.hpp
@@ -2,9 +2,10 @@
  * file_io.hpp
  */
 
+#include <cstdint>
 #include <cstdlib>
 
 char* loadFile( const char *fname, size_t &fsize );
 
-size_t getFilesize( const char *fname );
+uint64_t getFilesize( const char *fname );
